+++ wayflowcore/src/wayflowcore/conversation.py
@@ -3,6 +3,8 @@
 # This software is under the Apache License 2.0
 # (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0) or Universal Permissive License
 # (UPL) 1.0 (LICENSE-UPL or https://oss.oracle.com/licenses/upl), at your option.
+import logging
+import warnings
 from abc import abstractmethod
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any, AsyncIterable, Callable, Dict, List, Optional, Sequence
@@ -13,8 +15,9 @@ from wayflowcore.conversationalcomponent import ConversationalComponent
 from wayflowcore.executors._events.event import Event
 from wayflowcore.executors.executionstatus import (
     ExecutionStatus,
-    ToolExecutionConfirmationStatus,
+    FinishedStatus,
     ToolRequestStatus,
+    UserMessageRequestStatus,
 )
 from wayflowcore.messagelist import Message, MessageList
 from wayflowcore.planning import ExecutionPlan
@@ -27,6 +30,8 @@ if TYPE_CHECKING:
     from wayflowcore.models._requesthelpers import TaggedMessageChunkType
     from wayflowcore.tools import ToolResult
 
+
+logger = logging.getLogger(__name__)
 
 ContextProviderType = Callable[["Conversation"], Any]
 
@@ -80,6 +85,12 @@ class Conversation(DataclassComponent):
         The ``Execution`` status is returned by the Assistant and indicates if the assistant yielded,
         finished the conversation.
         """
+        if self.status is not None and self._status_applies_to_this_conversation(self.status):
+            self._update_conversation_with_status(self.status)
+
+        # reset status
+        self.status = None
+
         self.status = await self.component.runner.execute_async(self, execution_interrupts)
         return self.status
 
@@ -124,7 +135,14 @@ class Conversation(DataclassComponent):
         message:
             message to append.
         """
-        self.message_list.append_message(message)
+        if self.status is None:
+            self.message_list.append_message(message)
+        elif isinstance(self.status, UserMessageRequestStatus):
+            self.append_user_message(message.content)
+        elif isinstance(self.status, ToolRequestStatus) and message.tool_result is not None:
+            self.append_tool_result(message.tool_result)
+        else:
+            self.message_list.append_message(message)
 
     def append_agent_message(self, agent_input: str, is_error: bool = False) -> None:
         """Append a new message object of type ``MessageType.AGENT`` to the messages list.
@@ -134,7 +152,7 @@ class Conversation(DataclassComponent):
         agent_input:
             message to append.
         """
-        self.message_list.append_agent_message(agent_input, is_error)
+        self.message_list.append_agent_message(agent_input=agent_input, is_error=is_error)
 
     def append_user_message(self, user_input: str) -> None:
         """Append a new message object of type ``MessageType.USER`` to the messages list.
@@ -144,7 +162,24 @@ class Conversation(DataclassComponent):
         user_input:
             message to append.
         """
-        self.message_list.append_user_message(user_input)
+
+        from wayflowcore.executors.interrupts.executioninterrupt import InterruptedExecutionStatus
+
+        if self.status is None:
+            self.message_list.append_user_message(user_input)
+        elif isinstance(self.status, UserMessageRequestStatus):
+            self.status.submit_user_response(user_input)
+        elif isinstance(self.status, InterruptedExecutionStatus):
+            self.message_list.append_user_message(user_input)
+        elif isinstance(self.status, FinishedStatus):
+            warnings.warn(
+                "Conversation was finished but a new user message was appended.", UserWarning
+            )
+            self.message_list.append_user_message(user_input)
+        else:
+            raise ValueError(
+                f"Should not append tool results to the current conversation ({user_input}, {self.status})"
+            )
 
     def append_tool_result(self, tool_result: "ToolResult") -> None:
         """Append a new message object of type ``MessageType.TOOL_RESULT`` to the messages list.
@@ -154,18 +189,10 @@ class Conversation(DataclassComponent):
         tool_result:
             message to append.
         """
-        if (
-            isinstance(self.status, (ToolRequestStatus, ToolExecutionConfirmationStatus))
-            and self.status._conversation_id == self.id
-        ):
-            # we should append to this conversation message list only if
-            # it is the conversation the ToolRequestStatus/ToolExecutionConfirmationStatus originated from.
-            self.message_list.append_tool_result(tool_result)
-        else:
-            # otherwise, we need to propagate the tool_result to sub conversations
-            # to append the message to the conversation bound to the ToolRequestStatus/ToolExecutionConfirmationStatus
-            for conv in self._get_all_sub_conversations():
-                conv.append_tool_result(tool_result)
+        if self.status is None or not isinstance(self.status, ToolRequestStatus):
+            raise ValueError("internal error")
+
+        self.status.submit_tool_result(tool_result)
 
     async def _append_streaming_message(
         self,
@@ -173,3 +200,28 @@ class Conversation(DataclassComponent):
         extract_func: Optional[Callable[[Any], "TaggedMessageChunkType"]] = None,
     ) -> Message:
         return await self.message_list._append_streaming_message(stream, extract_func)
+
+    def _status_applies_to_this_conversation(self, status: ExecutionStatus) -> bool:
+        return status is not None and status._conversation_id == self.id
+
+    def _update_conversation_with_status(self, status: ExecutionStatus) -> None:
+        if isinstance(status, UserMessageRequestStatus):
+            user_response = status._user_response
+            if user_response is None:
+                s = "\n".join([str(m) for m in self.get_messages()])
+                logger.info(
+                    f"Should have submitted a user response before updating the conversation: {s}"
+                )
+            else:
+                self.message_list.append_message(user_response)
+        elif isinstance(status, ToolRequestStatus):
+            tool_results = status._tool_results
+            if tool_results is None:
+                # return
+                raise ValueError(
+                    "Should have submitted a tool results before updating the conversation\n"
+                    f"The status was: {status}\n"
+                    f"The conversation was: {self.get_messages()}\n"
+                )
+            for tool_result in tool_results:
+                self.message_list.append_message(Message(tool_result=tool_result, role="assistant"))